down to under 1MB
