+++ client/src/features/dashboard/components/DashboardMessage.tsx
@@ -16,13 +16,13 @@
  * limitations under the License.
  */
 
-import { useCallback, useContext, useMemo, useState } from "react";
 import cx from "classnames";
+import { useCallback, useContext, useMemo, useState } from "react";
 import { ChevronDown } from "react-bootstrap-icons";
 import { useDispatch } from "react-redux";
 import { Collapse } from "reactstrap";
 import { RenkuAlert } from "../../../components/Alert";
-import { RenkuMarkdown } from "../../../components/markdown/RenkuMarkdown";
+import LazyRenkuMarkdown from "../../../components/markdown/LazyRenkuMarkdown";
 import AppContext from "../../../utils/context/appContext";
 import { validateDashboardMessageParams } from "../message/dashboardMessage.utils";
 import {
@@ -65,7 +65,7 @@ export default function DashboardMessage() {
       dismissCallback={dismiss}
       dataCy="dashboard-message"
     >
-      <RenkuMarkdown markdownText={text} />
+      <LazyRenkuMarkdown markdownText={text} />
       <DashboardMessageMore additionalText={additionalText} />
     </RenkuAlert>
   );
@@ -98,7 +98,7 @@ function DashboardMessageMore({ additionalText }: { additionalText: string }) {
         </a>
       </p>
       <Collapse isOpen={show}>
-        <RenkuMarkdown markdownText={additionalText} />
+        <LazyRenkuMarkdown markdownText={additionalText} />
       </Collapse>
     </>
   );
