+++ client/src/utils/helpers/HelperFunctions.js
@@ -16,14 +16,6 @@
  * limitations under the License.
  */
 
-// title.Author: Alex K. - https://stackoverflow.com/users/246342/alex-k
-// Source: https://stackoverflow.com/questions/6507056/replace-all-whitespace-characters/6507078#6507078
-import showdown from "showdown";
-import showdownHighlight from "showdown-highlight";
-// Version 0.8.0 of showdown-katex breaks the tests so do not update the library
-import showdownKatex from "showdown-katex";
-import showdownMermaid from "showdown-mermaid";
-import DOMPurify from "dompurify";
 import XRegExp from "xregexp";
 
 const AUTOSAVED_PREFIX = "renku/autosave/";
@@ -155,64 +147,6 @@ function splitAutosavedBranches(branches) {
     (branch) => !branch.name.startsWith(AUTOSAVED_PREFIX)
   );
   return { standard, autosaved };
-}
-
-function sanitizedHTMLFromMarkdown(markdown, singleLine = false) {
-  // Reference: https://github.com/showdownjs/showdown/wiki/Showdown-Options
-  const showdownOptions = {
-    ghCompatibleHeaderId: true,
-    parseImgDimensions: true,
-    simplifiedAutoLink: true,
-    excludeTrailingPunctuationFromURLs: true,
-    strikethrough: true, // eslint-disable-line
-    tables: true,
-    tasklists: true, // eslint-disable-line
-    disableForced4SpacesIndentedSublists: true, // eslint-disable-line
-    literalMidWordUnderscores: true,
-    emoji: true,
-  };
-  const showdownClasses = {
-    table: "table word-break-normal",
-  };
-  // Reference: https://github.com/showdownjs/showdown/wiki/Add-default-classes-for-each-HTML-element
-  const bindings = Object.keys(showdownClasses).map((key) => ({
-    type: "output",
-    regex: new RegExp(`<${key}(.*?)(?:(class="([^"]*)")(.*))?>`, "g"),
-    replace: `<${key} $1 class="$3 ${showdownClasses[key]}" $4>`,
-  }));
-
-  const converter = new showdown.Converter({
-    ...showdownOptions,
-    extensions: [
-      ...bindings,
-      showdownMermaid(),
-      showdownHighlight({ pre: true }),
-      showdownKatex({
-        throwOnError: false,
-        displayMode: true,
-        errorColor: "var(--bs-danger)",
-        delimiters: [{ left: "$", right: "$", display: false }],
-      }),
-    ],
-  });
-  if (singleLine && markdown) {
-    const lineBreakers = ["<br>", "<br />", "<br/>", "\n"];
-    const breakPosition = Math.max(
-      ...lineBreakers.map((elem) => markdown.indexOf(elem))
-    );
-    if (breakPosition !== -1) markdown = markdown.substring(0, breakPosition);
-  }
-
-  // adding gitlab delimiters support https://docs.gitlab.com/ee/user/markdown.html#math
-  // inline math code between $` and `$
-  markdown = markdown?.replace(new RegExp(/\$`/i, "gm"), "$");
-  markdown = markdown?.replace(new RegExp(/`\$/i, "gm"), "$");
-  // Reference https://github.com/obedm503/showdown-katex
-  // this showdown extension only support ```ascii math or ```latex
-  markdown = markdown?.replace(new RegExp("\\```math", "gm"), "```latex");
-  const htmlFromMarkdown = converter.makeHtml(markdown);
-  const sanitized = DOMPurify.sanitize(htmlFromMarkdown);
-  return sanitized;
 }
 
 function simpleHash(str) {
@@ -438,7 +372,6 @@ export {
   slugFromTitle,
   getActiveProjectPathWithNamespace,
   splitAutosavedBranches,
-  sanitizedHTMLFromMarkdown,
   simpleHash,
   parseINIString,
   formatBytes,
